updates the eos_interface module
This commit updates the eos_interface module to use rbeapi.  It also modifies
the module parameters to be generally available for any interface in EOS.  It
supports the following parameters:

* enable
* description

The other parameters are being refactored into the eos_ethernet interface.
Merging this module could break existing eos_interface implementations