adds eos_ethernet module
This commit adds the eos_ethernet module to the set of available Puppet
modules.  The eos_ethernet module is designed for managing physical Ethernet
interfaces and supports the following parameters:

* name (namevar)
* description
* enable
* flowcontrol_send
* flwocontrol_receive

This commit also addes the unit tests for the eos_ethernet module.  This commit
should be generally save to apply to existing implementations