updates eos_switchport to use rbeapi
This commit updates the eos_switchport provider to use the rbeapi lib instead
of the puppet_x/eos library.  It also updates the unit tests for the
eos_switchport module and adds the test fixtures.  No other parameters have
been changed in this update.  This commit should be generally safe for use.