updates eos_portchannel type and provider
This commit updates the eos_portchannel type and provider to take advantage
of rbeapi.  It also includes fixtures and unit tests for both the updated
type and provider.

The eos_portchannel type adds a new parameter for configuration of
eos_portchannel resources:

* minimum_links

This commit should be generally applicable to all implementations of
eos_portchannel