fixes an issue creating eos_ethernet instances
This commit addresses an issue where the call to instances would generate
an error trying to add the flowcontrol parameters.  The instances method
now checks to be sure its an ethernet interface type before trying to
access the flowcontrol attributes.  This commit should be applied for proper
operation of eos_ethernet