updates the eos_system module to use rbeapi
This commit updates the eos_system module to use rbeapi instead of the modules
in puppet_x/eos.  This update includes unit tests and fixtures to support the
new type and provider.

This update changes how the eos_system module works. It now uses a namevar
of settings to identify the resource and adds a new parameter 'hostname' used
to configure the global hostname of the system.