updates the provider to support eos_ntp_config
This commit updates the provider to support the eos_ntp_config type. It includes
updated fixtures and unit spec tests.  This update is required to support the
latest rbeapi lib.

This commit also changes the dir name for the eos_ntp_server unit tests to be
correct.  No functional changes where introduced